move Indicator title as own component
